[update] Use async for read file
