+++ src/api/helpers/index.js
@@ -2,5 +2,6 @@ export default {
   MongooseHelper: require('./mongoose'),
   ResponseHelper: require('./response'),
   SendgridHelper: require('./sendgrid'),
-  TextureHelper: require('./texture')
+  TextureHelper: require('./texture'),
+  ModelHelper: require('./model')
 };
